[FLINK-30665][table] move some test to SqlOtherOperationConverterTest
