+++ flink-table/flink-table-planner/src/test/java/org/apache/flink/table/planner/operations/SqlDdlToOperationConverterTest.java
@@ -41,19 +41,10 @@ import org.apache.flink.table.catalog.exceptions.DatabaseNotExistException;
 import org.apache.flink.table.catalog.exceptions.FunctionAlreadyExistException;
 import org.apache.flink.table.expressions.SqlCallExpression;
 import org.apache.flink.table.factories.TestManagedTableFactory;
-import org.apache.flink.table.operations.LoadModuleOperation;
 import org.apache.flink.table.operations.NopOperation;
 import org.apache.flink.table.operations.Operation;
-import org.apache.flink.table.operations.ShowFunctionsOperation;
-import org.apache.flink.table.operations.ShowModulesOperation;
-import org.apache.flink.table.operations.ShowTablesOperation;
 import org.apache.flink.table.operations.SinkModifyOperation;
 import org.apache.flink.table.operations.SourceQueryOperation;
-import org.apache.flink.table.operations.UnloadModuleOperation;
-import org.apache.flink.table.operations.UseCatalogOperation;
-import org.apache.flink.table.operations.UseDatabaseOperation;
-import org.apache.flink.table.operations.UseModulesOperation;
-import org.apache.flink.table.operations.command.ShowJarsOperation;
 import org.apache.flink.table.operations.ddl.AlterDatabaseOperation;
 import org.apache.flink.table.operations.ddl.AlterTableChangeOperation;
 import org.apache.flink.table.operations.ddl.AlterTableRenameOperation;
@@ -99,36 +90,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /** Test cases for the DDL statements for {@link SqlToOperationConverter}. */
 public class SqlDdlToOperationConverterTest extends SqlToOperationConverterTestBase {
-
-    @Test
-    public void testUseCatalog() {
-        final String sql = "USE CATALOG cat1";
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(UseCatalogOperation.class);
-        assertThat(((UseCatalogOperation) operation).getCatalogName()).isEqualTo("cat1");
-        assertThat(operation.asSummaryString()).isEqualTo("USE CATALOG cat1");
-    }
-
-    @Test
-    public void testUseDatabase() {
-        final String sql1 = "USE db1";
-        Operation operation1 = parse(sql1);
-        assertThat(operation1).isInstanceOf(UseDatabaseOperation.class);
-        assertThat(((UseDatabaseOperation) operation1).getCatalogName()).isEqualTo("builtin");
-        assertThat(((UseDatabaseOperation) operation1).getDatabaseName()).isEqualTo("db1");
-
-        final String sql2 = "USE cat1.db1";
-        Operation operation2 = parse(sql2);
-        assertThat(operation2).isInstanceOf(UseDatabaseOperation.class);
-        assertThat(((UseDatabaseOperation) operation2).getCatalogName()).isEqualTo("cat1");
-        assertThat(((UseDatabaseOperation) operation2).getDatabaseName()).isEqualTo("db1");
-    }
-
-    @Test
-    public void testUseDatabaseWithException() {
-        final String sql = "USE cat1.db1.tbl1";
-        assertThatThrownBy(() -> parse(sql)).isInstanceOf(ValidationException.class);
-    }
 
     @Test
     public void testCreateDatabase() {
@@ -217,122 +178,6 @@ public class SqlDdlToOperationConverterTest extends SqlToOperationConverterTestB
                 .isEqualTo("db1_comment");
         assertThat(alterDatabaseOperation.getCatalogDatabase().getProperties())
                 .isEqualTo(properties);
-    }
-
-    @Test
-    public void testLoadModule() {
-        final String sql = "LOAD MODULE dummy WITH ('k1' = 'v1', 'k2' = 'v2')";
-        final String expectedModuleName = "dummy";
-        final Map<String, String> expectedOptions = new HashMap<>();
-        expectedOptions.put("k1", "v1");
-        expectedOptions.put("k2", "v2");
-
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(LoadModuleOperation.class);
-        final LoadModuleOperation loadModuleOperation = (LoadModuleOperation) operation;
-
-        assertThat(loadModuleOperation.getModuleName()).isEqualTo(expectedModuleName);
-        assertThat(loadModuleOperation.getOptions()).isEqualTo(expectedOptions);
-    }
-
-    @Test
-    public void testUnloadModule() {
-        final String sql = "UNLOAD MODULE dummy";
-        final String expectedModuleName = "dummy";
-
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(UnloadModuleOperation.class);
-
-        final UnloadModuleOperation unloadModuleOperation = (UnloadModuleOperation) operation;
-
-        assertThat(unloadModuleOperation.getModuleName()).isEqualTo(expectedModuleName);
-    }
-
-    @Test
-    public void testUseOneModule() {
-        final String sql = "USE MODULES dummy";
-        final List<String> expectedModuleNames = Collections.singletonList("dummy");
-
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(UseModulesOperation.class);
-
-        final UseModulesOperation useModulesOperation = (UseModulesOperation) operation;
-
-        assertThat(useModulesOperation.getModuleNames()).isEqualTo(expectedModuleNames);
-        assertThat(useModulesOperation.asSummaryString()).isEqualTo("USE MODULES: [dummy]");
-    }
-
-    @Test
-    public void testUseMultipleModules() {
-        final String sql = "USE MODULES x, y, z";
-        final List<String> expectedModuleNames = Arrays.asList("x", "y", "z");
-
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(UseModulesOperation.class);
-
-        final UseModulesOperation useModulesOperation = (UseModulesOperation) operation;
-
-        assertThat(useModulesOperation.getModuleNames()).isEqualTo(expectedModuleNames);
-        assertThat(useModulesOperation.asSummaryString()).isEqualTo("USE MODULES: [x, y, z]");
-    }
-
-    @Test
-    public void testShowModules() {
-        final String sql = "SHOW MODULES";
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(ShowModulesOperation.class);
-        final ShowModulesOperation showModulesOperation = (ShowModulesOperation) operation;
-
-        assertThat(showModulesOperation.requireFull()).isFalse();
-        assertThat(showModulesOperation.asSummaryString()).isEqualTo("SHOW MODULES");
-    }
-
-    @Test
-    public void testShowTables() {
-        final String sql = "SHOW TABLES from cat1.db1 not like 't%'";
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(ShowTablesOperation.class);
-
-        ShowTablesOperation showTablesOperation = (ShowTablesOperation) operation;
-        assertThat(showTablesOperation.getCatalogName()).isEqualTo("cat1");
-        assertThat(showTablesOperation.getDatabaseName()).isEqualTo("db1");
-        assertThat(showTablesOperation.getPreposition()).isEqualTo("FROM");
-        assertThat(showTablesOperation.isUseLike()).isTrue();
-        assertThat(showTablesOperation.isNotLike()).isTrue();
-
-        final String sql2 = "SHOW TABLES in db2";
-        showTablesOperation = (ShowTablesOperation) parse(sql2);
-        assertThat(showTablesOperation.getCatalogName()).isEqualTo("builtin");
-        assertThat(showTablesOperation.getDatabaseName()).isEqualTo("db2");
-        assertThat(showTablesOperation.getPreposition()).isEqualTo("IN");
-        assertThat(showTablesOperation.isUseLike()).isFalse();
-        assertThat(showTablesOperation.isNotLike()).isFalse();
-
-        final String sql3 = "SHOW TABLES";
-        showTablesOperation = (ShowTablesOperation) parse(sql3);
-        assertThat(showTablesOperation.getCatalogName()).isNull();
-        assertThat(showTablesOperation.getDatabaseName()).isNull();
-        assertThat(showTablesOperation.getPreposition()).isNull();
-    }
-
-    @Test
-    public void testShowFullModules() {
-        final String sql = "SHOW FULL MODULES";
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(ShowModulesOperation.class);
-        final ShowModulesOperation showModulesOperation = (ShowModulesOperation) operation;
-
-        assertThat(showModulesOperation.requireFull()).isTrue();
-        assertThat(showModulesOperation.asSummaryString()).isEqualTo("SHOW FULL MODULES");
-    }
-
-    @Test
-    public void testShowFunctions() {
-        final String sql1 = "SHOW FUNCTIONS";
-        assertShowFunctions(sql1, sql1, ShowFunctionsOperation.FunctionScope.ALL);
-
-        final String sql2 = "SHOW USER FUNCTIONS";
-        assertShowFunctions(sql2, sql2, ShowFunctionsOperation.FunctionScope.USER);
     }
 
     @Test
@@ -2304,15 +2149,6 @@ public class SqlDdlToOperationConverterTest extends SqlToOperationConverterTestB
         assertThat(operation).isInstanceOf(CreateViewOperation.class);
     }
 
-    @Test
-    public void testShowJars() {
-        final String sql = "SHOW JARS";
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(ShowJarsOperation.class);
-        final ShowJarsOperation showModulesOperation = (ShowJarsOperation) operation;
-        assertThat(showModulesOperation.asSummaryString()).isEqualTo("SHOW JARS");
-    }
-
     // ~ Tool Methods ----------------------------------------------------------
 
     private static TestItem createTestItem(Object... args) {
@@ -2407,19 +2243,6 @@ public class SqlDdlToOperationConverterTest extends SqlToOperationConverterTestB
         catalogManager.setCurrentDatabase("db1");
         ObjectIdentifier tableIdentifier = ObjectIdentifier.of("cat1", "db1", tableName);
         catalogManager.createTable(catalogTable, tableIdentifier, true);
-    }
-
-    private void assertShowFunctions(
-            String sql,
-            String expectedSummary,
-            ShowFunctionsOperation.FunctionScope expectedScope) {
-        Operation operation = parse(sql);
-        assertThat(operation).isInstanceOf(ShowFunctionsOperation.class);
-
-        final ShowFunctionsOperation showFunctionsOperation = (ShowFunctionsOperation) operation;
-
-        assertThat(showFunctionsOperation.getFunctionScope()).isEqualTo(expectedScope);
-        assertThat(showFunctionsOperation.asSummaryString()).isEqualTo(expectedSummary);
     }
 
     private void assertAlterTableOptions(
